+++ src/util/append-or-not.ts
@@ -1,3 +1,0 @@
-export function appendOrNot(count: number, singular: string): string {
-    return `${ count } ${singular}${ count === 1 ? "" : "s" } ago`;
-}