+++ src/util/message.ts
@@ -2,10 +2,10 @@ import { MessageItem, window } from "vscode";
 
 export const infoMessage = <T extends MessageItem>(
     message: string,
-    ...items: T[]
-): Promise<T | undefined> => Promise.resolve(
-    window.showInformationMessage(message, ...items),
-);
+    item: undefined | T[] = [],
+): Promise<T | undefined> => {
+    return Promise.resolve(window.showInformationMessage(message, ...item));
+}
 
 export const errorMessage = (
     message: string,