+++ src/util/property.ts
@@ -13,17 +13,8 @@ export type PropertiesMap = {
     "pluralWebPathSubstrings": string[];
 }
 
-type PropertyFunction = {
-    <Key extends keyof PropertiesMap>(
-        name: Key,
-        fallback: PropertiesMap[Key],
-    ): PropertiesMap[Key];
-    <Key extends keyof PropertiesMap>(
-        name: Key,
-    ): PropertiesMap[Key] | undefined;
-}
-
-export const getProperty: PropertyFunction = <Key extends keyof PropertiesMap>(
+// getConfiguration has an unfortunate typing that does not
+// take any possible default values into consideration.
+export const getProperty = <Key extends keyof PropertiesMap>(
     name: Key,
-    fallback?: PropertiesMap[Key],
-): PropertiesMap[Key] | undefined => workspace.getConfiguration(extensionName).get(name) ?? fallback;
+): PropertiesMap[Key] => workspace.getConfiguration(extensionName).get(name) as PropertiesMap[Key];