+++ src/view.ts
@@ -22,7 +22,7 @@ export class StatusBarView {
         if (!commit) {
             this.setText("", false);
         } else if (isUncomitted(commit)) {
-            this.setText(getProperty("statusBarMessageNoCommit", "Not Committed Yet"), false);
+            this.setText(getProperty("statusBarMessageNoCommit"), false);
         } else {
             this.setText(toTextView(commit), true);
         }
@@ -37,9 +37,8 @@ export class StatusBarView {
     }
 
     private setText(text: string, command: boolean): void {
-        const noInfo = " - No info about the current line";
         this.out.text = "$(git-commit) " + text.trimEnd();
-        this.out.tooltip = `git blame${ command ? "" : noInfo }`;
+        this.out.tooltip = `git blame${ command ? "" : " - No info about the current line" }`;
         this.out.command = command ? `${extensionName}.quickInfo` : undefined;
     }
 }