+++ test/suite/textdecorator.test.ts
@@ -2,24 +2,24 @@ import * as assert from "assert";
 import { useFakeTimers } from "sinon";
 
 import { Commit } from "../../src/git/util/stream-parsing";
+import { between } from "../../src/util/ago";
 import {
     InfoTokens,
     normalizeCommitInfoTokens,
     parseTokens,
-    toDateText,
 } from "../../src/util/textdecorator";
 
 suite("Date Calculations", (): void => {
     test("Time ago in years", (): void => {
         assert.strictEqual(
-            toDateText(
-                new Date(2015, 1),
+            between(
+                new Date(2015, 2),
                 new Date(2014, 1),
             ),
             "1 year ago",
         );
         assert.strictEqual(
-            toDateText(
+            between(
                 new Date(2015, 1),
                 new Date(2005, 1),
             ),
@@ -29,31 +29,38 @@ suite("Date Calculations", (): void => {
 
     test("Time ago in months", (): void => {
         assert.strictEqual(
-            toDateText(
+            between(
                 new Date(2015, 1),
                 new Date(2015, 0),
             ),
             "1 month ago",
         );
         assert.strictEqual(
-            toDateText(
+            between(
                 new Date(2015, 11, 10),
                 new Date(2015, 0),
             ),
             "11 months ago",
         );
+        assert.strictEqual(
+            between(
+                new Date(2015, 1),
+                new Date(2014, 1),
+            ),
+            "12 months ago",
+        );
     });
 
     test("Time ago in days", (): void => {
         assert.strictEqual(
-            toDateText(
-                new Date(2015, 1, 2),
-                new Date(2015, 1, 1),
+            between(
+                new Date(2015, 1, 2, 8),
+                new Date(2015, 1, 1, 0),
             ),
             "1 day ago",
         );
         assert.strictEqual(
-            toDateText(
+            between(
                 new Date(2015, 1, 31),
                 new Date(2015, 1, 1),
             ),
@@ -63,49 +70,56 @@ suite("Date Calculations", (): void => {
 
     test("Time ago in hours", (): void => {
         assert.strictEqual(
-            toDateText(
-                new Date(2015, 1, 1, 1, 0, 0),
+            between(
+                new Date(2015, 1, 1, 1, 5, 0),
                 new Date(2015, 1, 1, 0, 0, 0),
             ),
             "1 hour ago",
         );
         assert.strictEqual(
-            toDateText(
+            between(
                 new Date(2015, 1, 1, 23, 29, 0),
                 new Date(2015, 1, 1, 0, 0, 0),
             ),
             "23 hours ago",
         );
+        assert.strictEqual(
+            between(
+                new Date(2015, 1, 2),
+                new Date(2015, 1, 1),
+            ),
+            "24 hours ago",
+        );
     });
 
     test("Time ago in minutes", (): void => {
         assert.strictEqual(
-            toDateText(
+            between(
                 new Date(2015, 1, 1, 1, 5, 0),
                 new Date(2015, 1, 1, 1, 0, 0),
             ),
             "5 minutes ago",
         );
         assert.strictEqual(
-            toDateText(
+            between(
                 new Date(2015, 1, 1, 1, 59, 29),
                 new Date(2015, 1, 1, 1, 0, 0),
             ),
             "59 minutes ago",
         );
+        assert.strictEqual(
+            between(
+                new Date(2015, 1, 1, 1, 0, 0),
+                new Date(2015, 1, 1, 0, 0, 0),
+            ),
+            "60 minutes ago",
+        );
     });
 
     test("Right now", (): void => {
         assert.strictEqual(
-            toDateText(
+            between(
                 new Date(2015, 1, 1, 1, 0, 1),
-                new Date(2015, 1, 1, 1, 0, 0),
-            ),
-            "right now",
-        );
-        assert.strictEqual(
-            toDateText(
-                new Date(2015, 1, 1, 1, 4, 29),
                 new Date(2015, 1, 1, 1, 0, 0),
             ),
             "right now",
@@ -279,6 +293,44 @@ suite("Token Parser", (): void => {
             "Multiple example-token in a length text",
         );
     });
+
+    test("Dangling token", (): void => {
+        assert.strictEqual(
+            parseTokens(
+                "This token ${is.not.closed",
+                normalizedInfo,
+            ),
+            "This token ${is.not.closed",
+        );
+        assert.strictEqual(
+            parseTokens(
+                "This token ${is.not.closed with spaces",
+                normalizedInfo,
+            ),
+            "This token ${is.not.closed with spaces",
+        );
+        assert.strictEqual(
+            parseTokens(
+                "This token ${is.not.closed,with_params",
+                normalizedInfo,
+            ),
+            "This token ${is.not.closed,with_params",
+        );
+        assert.strictEqual(
+            parseTokens(
+                "This token ${is.not.closed|with_modifier",
+                normalizedInfo,
+            ),
+            "This token ${is.not.closed|with_modifier",
+        );
+        assert.strictEqual(
+            parseTokens(
+                "This token ${is.not.closed,with_params|and_modifier",
+                normalizedInfo,
+            ),
+            "This token ${is.not.closed,with_params|and_modifier",
+        );
+    });
 });
 
 suite("Text Decorator with CommitInfoToken", (): void => {
@@ -287,13 +339,15 @@ suite("Text Decorator with CommitInfoToken", (): void => {
         "author": {
             "mail": "<vdavydov.dev@gmail.com>",
             "name": "Vladimir Davydov",
-            "timestamp": 1423781950,
+            "timestamp": "1423781950",
+            "date": new Date(1_423_781_950_000),
             "tz": "-0800",
         },
         "committer": {
             "mail": "<torvalds@linux-foundation.org>",
             "name": "Linus Torvalds",
-            "timestamp": 1423796049,
+            "timestamp": "1423796049",
+            "date": new Date(1_423_796_049_000),
             "tz": "-0800",
         },
         "hash": "60d3fd32a7a9da4c8c93a9f89cfda22a0b4c65ce",
@@ -312,6 +366,8 @@ suite("Text Decorator with CommitInfoToken", (): void => {
 
     check("author.mail", "<vdavydov.dev@gmail.com>");
     check("author.name", "Vladimir Davydov");
+    check("author.name|u", "VLADIMIR DAVYDOV");
+    check("author.name|l", "vladimir davydov");
     check("author.tz", "-0800");
     check("author.date", "2015-02-12");
 
@@ -326,12 +382,12 @@ suite("Text Decorator with CommitInfoToken", (): void => {
 
     check("time.ago", "6 years ago");
     check("time.c_ago", "6 years ago");
-    check("time.from", "6 years ago");
-    check("time.c_from", "6 years ago");
 
     check("commit.summary,0", "");
     check("commit.summary,5", "list_");
+    check("commit.summary,5|u", "LIST_");
     check("commit.hash_short,0", "");
     check("commit.hash_short,2", "60");
     check("commit.hash_short,39", "60d3fd32a7a9da4c8c93a9f89cfda22a0b4c65c");
+    check("commit.hash_short,100|u", "60D3FD32A7A9DA4C8C93A9F89CFDA22A0B4C65CE");
 });